Add initial project setup
This commit includes the initial setup for a Kotlin project. The following .xml files have been added to the .idea directory which include settings for the Kotlin plugin and Miscellanous project settings: kotlinc.xml, inspectionProfiles/Project_Default.xml, misc.xml, gradle.xml and vcs.xml. It also introduces the Gradle wrapper for building the project along with the gradlew.bat and gradlew scripts for running on Windows and Unix respectively. Adds files: build.gradle.kts for the build settings, gradle.properties to specify the code style and settings.gradle.kts for project settings. Moreover, the .gitignore file was added in both root and .idea directory for ignoring unnecessary files. Changes are done to set up the project environment and enforce coding standards.