Disable DuplicatedCode Inspection in Project Default Profile
An update has been made in the inspectionProfiles where the 'DuplicatedCode' tool was added and set to 'disabled'. This is to prevent warning prompts related to duplicated code which may be intentional and not necessarily erroneous in the coding structure. This also gives a less cluttered warning environment for the developer.