Remove Calibration Calculations and Introduce CalibrationLine classes
Deleted the separate Calibration files that individually handled each part of the calibration process. Instead, Introduced a new class 'CalibrationLine' that encapsulates the logic for both basic and complex calibrations, bringing them into a singular scope. Also, 'TrebuchetExecutor' is introduced as the new entry point for execution. These changes aimed to simplify and make the codebase more maintainable. Also, Unit-tests are adapted to the new changes and are passing successfully.